Complet ProductVariantItem

+++ app/Http/Controllers/Backend/ProductVariantController.php
@@ -6,6 +6,7 @@ use App\DataTables\ProductVariantDataTable;
 use App\Http\Controllers\Controller;
 use App\Models\Product;
 use App\Models\ProductVariant;
+use App\Models\ProductVariantItem;
 use Illuminate\Http\Request;
 
 class ProductVariantController extends Controller
@@ -89,6 +90,10 @@ class ProductVariantController extends Controller
     public function destroy(string $id)
     {
         $variant = ProductVariant::findOrFail($id);
+        $variantItem = ProductVariantItem::where('product_variant_id', $variant->id)->count();
+        if ($variantItem > 0) {
+            return response(['status' => 'error', 'message' => 'Cannot delete this variant, it has items.']);
+        }
         $variant->delete();
         return response(['status' => 'success', 'message' => 'Deleted Successfully!']);
     }
